DB컬럼의 자료형에 맞게 전달되는 파라미터의 자료형도 변경

+++ src/main/java/com/semi/easycoding/education/mapper/EducationMapper.java
@@ -35,7 +35,7 @@ public interface EducationMapper {
     List<EducationDto> selectEducationNotAssigned(@Param("memberId") Long memberId);
     // 사용자에게 할당되지 않은 문제 조회
     List<EducationDto> selectEducationNotAssignedByCategory(@Param("memberId") Long memberId,
-                                                            @Param("categoryId") Long categoryId);
+                                                            @Param("categoryId") Short categoryId);
 
 
     int insertMemberQuizHistory(@Param("memberId") Long memberId,
